+++ src/main/java/com/example/b_team_shopping_mall/dto/Claim/ClaimCreateRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.b_team_shopping_mall.dto;
+package com.example.b_team_shopping_mall.dto.Claim;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -8,12 +8,12 @@ import javax.validation.constraints.NotBlank;
 @AllArgsConstructor
 @Data
 public class ClaimCreateRequestDto {
+
+    private String memberid;
+
     @NotBlank(message = "문의하실 사항의 제목을 입력해주세요")
     private String title;
 
     @NotBlank(message = "문의하실 내용을 입력해주세요")
     private String content;
-
-    @NotBlank(message = "문의하시는 분의 아이디를 입력해주세요")
-    private String writer;
 }
