+++ src/main/java/com/example/b_team_shopping_mall/dto/Claim/ClaimCreateResponseDto.java
@@ -1,0 +1,19 @@
+package com.example.b_team_shopping_mall.dto.Claim;
+
+import com.example.b_team_shopping_mall.entity.Claim;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class ClaimCreateResponseDto {
+    private String title;
+    private String content;
+    private String writer;
+
+    public ClaimCreateResponseDto toDto(Claim claim) {
+        return new ClaimCreateResponseDto(claim.getTitle(), claim.getContent(), claim.getRegister().getMembername());
+    }
+}
