+++ src/main/java/com/example/b_team_shopping_mall/dto/Claim/ClaimGetResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.b_team_shopping_mall.dto;
+package com.example.b_team_shopping_mall.dto.Claim;
 
 import com.example.b_team_shopping_mall.entity.Claim;
 import lombok.AllArgsConstructor;
@@ -14,6 +14,6 @@ public class ClaimGetResponseDto {
     private String writer;
 
     public ClaimGetResponseDto toDto(Claim claim) {
-        return new ClaimGetResponseDto(claim.getTitle(), claim.getContent(), claim.getWriter());
+        return new ClaimGetResponseDto(claim.getTitle(), claim.getContent(), claim.getRegister().getMembername());
     }
 }
