+++ src/main/java/com/example/b_team_shopping_mall/dto/Claim/ClaimListResponseDto.java
@@ -1,19 +1,19 @@
-package com.example.b_team_shopping_mall.dto;
+package com.example.b_team_shopping_mall.dto.Claim;
 
 import com.example.b_team_shopping_mall.entity.Claim;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-@Data
 @AllArgsConstructor
+@Data
 @NoArgsConstructor
-public class ClaimCreateResponseDto {
+public class ClaimListResponseDto {
     private String title;
     private String content;
     private String writer;
 
-    public ClaimCreateResponseDto toDto(Claim claim) {
-        return new ClaimCreateResponseDto(claim.getTitle(), claim.getContent(), claim.getWriter());
+    public ClaimListResponseDto toDto(Claim claim) {
+        return new ClaimListResponseDto(claim.getTitle(), claim.getContent(), claim.getRegister().getMembername());
     }
 }
